backend/requirements.txt
flask
flask_cors
yfinance
twilio
python-dotenv
requests